♻️ Decompose store listing page
Extract the tag system into useStoreTags, and the search modal,
affiliate CTA / Plus welcome banners (with their scoped styles), and
list empty/loading states into components.
